+++ NEWS.md
@@ -3,6 +3,10 @@
 ## New Features
 
 * target_dir parameter accepts a special value 'stdout' to print to console. This is useful for 'piping' in bioinformatics applications
+* extracts now all contigs in an .ace file
+
+## Changes
+* returned object is of type list (instead of character)
 
 ## Fixes